requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
greenlet>=3.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
cryptography>=42.0.0  # For Fernet encryption of API keys

# HTTP Client (for integrations)
httpx>=0.26.0
aiohttp>=3.9.0

# WebSockets (for real-time canvas)
websockets>=12.0

# Utilities
python-dateutil>=2.8.0
orjson>=3.9.0

# MCP Server
mcp>=1.0.0

# Development
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.2.0
